// rea_frontend/app/(dashboard)/help-center/models/types/HelpCenter.ts
import { Icon } from "@tabler/icons-react";

export type ContactOption = {
    icon: Icon;
    title: string;
    description: string;
    contact: string;
    color: string;
}

export type FaqItem = {
    question: string;
    answer: string;
}

export type Resource = {
    icon: Icon;
    title: string;
    description: string;
}